move forward declaration header, enable hashing empty range
