+++ include/yk/hash/hash_value_for/fwd.hpp
